Update Block Broken Windows updates by KBID
M$ will push the update into "downloading" before we can run the script, which will prevent it from working as expected. I was able to reproduce the issue where once the update is in downloading status, blocking it in this manner does not work consistently. This update will partially handle that. Make sure you boot into safe mode with networking enabled to properly run this file. Since safe mode escalates the script as an administrator, you shouldn't need to run this as an admin. I'd still recommend using the "Windows Update Blocker" to see if you can block the offending update that way. I will be adding new KB updates that are problematic to this list as I come across them.