Create wine32
You don't actually need a .sh extension to make scripts work. Some people like them, some are okay without. I usually don't have the .sh extension, but it depends on the script. In this case, I go without. I also wrote this because some people were talking about how difficult it is to switch between wine prefixes, so this isn't needed if you are already familiar with how, but I like to have a normal wine prefix for 64-bit apps, and this wine32 just for the oddball exception.