Create CopySpotlightImages.bat
I contemplated whether I wanted to upload this since this is so specific, but hey, it contains a few tricks I've picked up and haven't yet posted here. So here you go. This creates a cache folder and copies all the files off the cache folder into the cache folder. Yikes that sounded weird. Anyways, it appears as though more than just wallpapers are stored in the cache folder, so you have both the wide and tall version of each wallpaper, along with the other junk that exists in the folder. I think the filenames are file hashes, which may explain the lack of file extension, and also means that we can safely ignore files that already exists in the cache folder, hence the "n" being piped into the copy command - it can automatically say "no" to overwriting files, since it isn't necessary. I have not yet decided how to handle renaming the files and putting them to use, so they just kind of sit in the cache folder, but I run it on computers from time to time to see if I can build a complete Spotlight image cache. Maybe one day, we will try and see if we can crowd source this project.