Create Music Genre Wheel of Doom.html
This little HTML page is what I like to call the "Genre Wheel of Doom", and it allows you to generate random genres, sub-genres, modifiers, vignettes, and a sample name, and use that for whatever madness your mind may carry. And it has a random number generator, because why not?

Thank you Echo, Gemini, and Copilot for helping out!