Create DisableDefender.ps1
A script that Copilot wrote for me that tries to disable Defender. It kind of works. Not a fault of Copilot, more of a fault of mine for not liking PowerShell and wishing for its demise. You will need to manually disable tamper protection and you will need to reboot, but once you turn off real time scanning, you can run other Windows Defender disabling scripts or programs to actually remove Defender. Or at least mitigate its teeth until the next Windows Update that renders it unusable.